+++ lib/views/course_view/components/base_card.dart
