+++ lib/views/course_view/downloaded_courses_view/download_card.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:gocast_mobile/views/course_view/components/baseCard.dart';
+import 'package:gocast_mobile/views/course_view/components/base_card.dart';
 
 class VideoCard extends BaseCard {
   final String title;
